requirements.txt
# Multi-Agent Inventory Management System Dependencies

# Core framework
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data handling
pandas>=2.0.0
requests>=2.31.0

# Google Sheets integration
gspread>=6.0.0
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0

# Math and calculations
numpy>=1.24.0